tag tree is now 64 bit

+++ src/lib/openjp2/t2.cpp
@@ -737,8 +737,8 @@ namespace grk {
 
 				/* if cblk not yet included before --> inclusion tagtree */
 				if (!l_cblk->numSegments) {
-					uint32_t value;
-					if (!l_prc->incltree->decodeValue(l_bio.get(), cblkno, (int32_t)(p_pi->layno + 1), &value)) {
+					uint64_t value;
+					if (!l_prc->incltree->decodeValue(l_bio.get(), cblkno, p_pi->layno + 1, &value)) {
 						event_msg(p_manager, EVT_ERROR, "t2_read_packet_header: failed to read `inclusion` bit \n");
 						return false;
 					}
@@ -772,10 +772,10 @@ namespace grk {
 
 				/* if cblk not yet included --> zero-bitplane tagtree */
 				if (!l_cblk->numSegments) {
-					uint32_t i = 0;
+					uint64_t i = 0;
 					uint8_t value;
 					bool rc = true;
-					while ( (rc = l_prc->imsbtree->decode(l_bio.get(), cblkno, (int32_t)i, &value)) && !value) {
+					while ( (rc = l_prc->imsbtree->decode(l_bio.get(), cblkno, i, &value)) && !value) {
 						++i;
 					}
 					if (!rc) {
@@ -783,7 +783,7 @@ namespace grk {
 						return false;
 					}
 
-					l_cblk->numbps = l_band->numbps + 1 - i;
+					l_cblk->numbps = l_band->numbps + 1 - (uint32_t)i;
 					// BIBO analysis gives upper limit on number of bit planes
 					if (l_cblk->numbps > max_precision_jpeg_2000 + OPJ_J2K_MAXRLVLS * 5) {
 						event_msg(p_manager, EVT_WARNING, "Number of bit planes %u is impossibly large.\n", l_cblk->numbps);
